PulseAudio not used.
So when doing the cleanup and verifying that the code actually works,
I found that even when you compile defimulator with PulseAudio and
PulseAudio-Simple support, it's not an option to be selected in the
advanced settings.  I'm not sure if it's just not included in the
settings portion of the GUI, or if it is legitimately turned off
somehow.  I think ALSA and OSS support will do all I need it to on
POSIX-like systems, and PulseAudio has a bit of a bad name anway.  I
will still probably move forward with PortAudio, however.