Hotkey Configuration added to MenuBar
Pretty hacky solution, but does the trick.  Launches the InputSettings
class with the proper device selected so that you don't have to
explore the menu to figure it out.

Also removed the 'unaquire input' functionality from the hotkeys.
I'm not sure what that does, so remove it for now.  I can always
bring it back with git.