Phoenix will not be used after the transition to GTK3.
Phoenix was built for bsnes.  But it's growing long in the tooth, and the
widget placement ("pixel-perfect precision") is terribly bad for any toolkit
that gets theme.  With the transition to GTK3, I hope to just make GTKMM3
calls from within the 'ui' portion of the project, without calling phoenix
which then calls the underlying toolkit.  GTKMM support looks like it's
pretty good on windows as well.

GTK3 will be stable soon.  But phoenix has to go.