Initial gtkmm-3.0 window is built.
Nothing works.  There is another main.cpp file created in the "test"
folder inside the ui-gtkmm3 folder that I'm using to open the window
with.  It's simply there to build and troubleshoot the window without
actually hooking up the emulation portion.

All the old UI code (that was dependent upon Phoenix) was removed.
It's a completely different paradigm with gtkmm-3.0, so I'm hesitant
to try and build upon the code that used phoenix.  I've still got the
code inside the ui/ and phoenix/ folders to fall back on if things
get rocky.