Adding Cartridge and FileBrowser support.
The signals and slots are still not hooked up, so that needs to be
done.  A lot of the GUI that you can see is just a shell of what
needs to be implemented.

A lot of cartridge code was redundant.  So I combined a lot of the
code into a single function with a switch statement.  If you look
in the Cartridge::load function, you'll see all the cartridge types
combined in there.

The FileBrowser class only has support for loading cartridges. If
there wasn't different slotted cartridges, I would probably just
use GTK's implementation of the file browser.  But because a lot of
the cartridges need to be loaded in two or three, a dialog has to
be created.  But support for the Pixel Shader and Video Filters
were removed from the FileBrowser.