Starting the InputMapper.
Which is easily the biggest mess of all the GUI code.  So I've got
that going for me, which is nice.

The hardest part will be that InputMapper refers to inputMapper,
which is assumes is a singleton of itself.  Which is fine, except
there are lots of variables that seem to fall from the sky when
untangling the code.