InputSettings is my first big obstacle.
Lots of global state will be introduced as a result.  The way the
input behaves is very dependent upon user configuration, window being
focused or emulation being paused.  Which is not something that can
a strict object of a certain class.

Before I move forward here, I'll create the input-configuration
dialog that the input mapper references.  I also need to work on
synchronizing the actual configuration class with the config dialogs.
Otherwise, it's all a waste.